Move SchemaComponentData to submodule
